Update patch set 12
Patch Set 12:

The percona method now works, although it would be nice to print the debug information about that, both on success and on error (I think percona has a verbose option (or maybe it is ignored), but definitely the errors should be printed too.

Very nice work, this is 99% (almost) there. Thank you!

Patch-set: 12