Update patch set 5
Patch Set 5:

So I got a full snapshot cycle- It took a long time, with only one failure- the latest backup wasn't archived to archive due to checking it was a directory (and not a tar.gz). My latest amend should fix that.

There are more hidden bugs but this should make this patch mostly feature-complete (for the description scope). I may deploy it and continue working on a separate one next week.

Patch-set: 5
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>