Update patch set 6
Patch Set 6:

I like the idea of being able to provide a given path, but I am not understanding what is the idea behind it, to be more precise this:
"This will prevent failures to cascade by indicating which
specific backup wants to be prepared, rather than failing because
it is ambiguous (there is a failed and a new one)."

Can you elaborate a bit on the purpose behind this? (Again, I do see the value and flexibility of being able to provide a given path rather than just using the default) but I want to make sure I fully understand what you meant behind that sentence.
Thanks!

Patch-set: 6
Reviewer: Marostegui <3731@e9e9afe9-4712-486d-8885-f54b72dd1951>