Update patch set 9
Patch Set 9: Code-Review+2

> Patch Set 6: Code-Review+1
> 
> Nice work! My only comment as I said the other day is that it would be more helpful to have all the hosts in the same DC with the same "+" colour.
> In the example you provide, all the hosts in codfw would have the pink "+" whereas the ones in eqiad would have the blue "+"

That's not easy- it requires a strict (code) definition to what "hosts in the same DC" is, and right now a host can only compare to its parent. Maybe what you are looking for is a `--dc=eqiad` and that can be done, but on a separate patch.

Merging this for now as this is stable, although a bit slow (again, more will be fixed later).

I am thinking of adding the master of the given instance, or allowing to pass a section (e.g. replication_tree.py s1), but all those may only be done later. There are non-trivial questions to the proposed extensions.

Patch-set: 9
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2