Update patch set 1
Patch Set 1:

Hi, Guozr.im,

Thanks for identifying the issue and sending a patch. I will temporarily assign the ticket to you to reflect you are working right now on it (so there is no duplicate work).

I have a couple of comments, let me know what you think about them:

As a solution, you capture the exception on the single line it can occur (that's great) and then abort the execution.

My first comment is that an else statement is not needed here because it increases cyclomatic complexity of the final block of code, without much need. I am very much in the "if (error) exit()", with no else needed, side. If you are going to abort early, there is no need for an else, as we know the rest will not be executed anyway. Too much "nesting" is confusing for reading code, and while we normally would solve this by creating extra functions, in this case we can just not use else and leave not nested the "most common code path" (when we do not capture the exception). Specially here makes sense as it avoids touching unnecessary extra lines of code, so limits the patch size.

My second comment is about the solution chosen- sys.exit. While this makes total sense in the context of a command line utility, note this is a library that could be used (in addition to transfer.py command line) as part of a multi-threaded program or a continuously running daemon (this is not theoretical, it is actually used like that). That is why it doesn't interact with the os library or tries to manage the flow of the program. I think it would be better if we just returned control to the caller and signaled an error in this case rather than trying to fully abort the program, which we should probably not control.

This means either returning with a failed command execution (one could argue that the execution did actually fail) and passing the error message as stder (third parameter) or a more elegant- capturing the exception, and rising a new exception for the caller to handle regarding a "invalid host" or something. Despite the last one being more elegant, to avoid increased complexity, I would suggest to keep it simple and just do the "return an error code different from 0 + custom error message (which could contain the text of the original exception)".

Just to be clear- this is not required to be completed for GSoC, I am already happy with you identifying the issue and sending a correctly formatted patch. This is in case you want it merged into production, as we would do for other contributors, and for learning purposes. If this was part of an ongoing GSoC (it isn't) i would also ask for a test case added with an invalid host to verify the bug is fixed (under the wmfmariadbpy/test dir).

I hope my comments would be useful to you anyway. Thank you again for your contribution, understanding and modifying this code is not an easy task.

Patch-set: 1
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>