Update patch set 6
Patch Set 6:

The only thing I have in mind at the moment is...given that this script is very specific for es (ie: will only work with single PKs, which are not the case in the rest of MW core tables), should it maybe be named in a different way? To specify it is just for es?

Patch-set: 6
Reviewer: Marostegui <3731@e9e9afe9-4712-486d-8885-f54b72dd1951>