Update patch set 2
Patch Set 2:

I hope it was clear that my purpose here was not to annoy you with red tape 0:-D

As a mentor, I have to demand following style standards, mostly for teaching purposes. These are not meaningless, those reviewing changes afterwards will thank you for a clear commit message, when/if problems arise, be them related or unrelated.

Thanks for your work, it is highly appreciated by the DBA team!

Patch-set: 2