Update patch set 3
Patch Set 3:

So Volans is our expert on Cumin, added here, but I think the best way to add a test here for our execution method is to mock the run command on Cumin (not trying to actually run it because it is not properly setup, and return an exception manually there (we probably don't want to fully configure Cumin for a unit test), and see if our execution class captures it appropriately and build the right CommandExecution object.

Volans: agree?

Patch-set: 3
Reviewer: Volans <3215@e9e9afe9-4712-486d-8885-f54b72dd1951>
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>