Update patch set 6
Patch Set 6:

Sorry, this is indeed fixed, only the line fix is needed for +2. I tested it on a different branch and works as expected.

It hit another unrelated bug:
./transfer.py test:/ test:/
ERROR: The specified source path / doesn't exist on test
Traceback (most recent call last):
  File "./transfer.py", line 511, in <module>
    main()
  File "./transfer.py", line 507, in main
    sys.exit(max(result))
TypeError: 'int' object is not iterable

But that would be a different patch, and it is not on scope of the ticket.

Patch-set: 6
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>