Update patch set 4
Patch Set 4:

I am a bit doubtful about 2 things: API and implementation.

Regarding API, sadly we cannot change unless there is no other way, the meaning of existing parameters- we should be backwards compatible as much as possible. This is because we cannot guaranteed that all applications that use this change at the same time
than the meaning change of the application. Does that mean that we cannot change the API? Yes, we can, but as much as possible,
only through adding new parameters, not modifying existing ones.

In the current patch, you modify the meaning of the port option, from "use this port" to "search for an open port starting from this".
We should not modify this behaviour and just error our if explicitly a port is required. For example, we could default to port 0 (or -1), and if a user doesn't use a particular port (defaults to 0 or -1), choose it ourselves.
We can start searching from the default port to a range (e.g. 4444-4499), and if you want to make that configurable (not required), we can add an aditional option "--auto-port-range" or something like that, but configurability is not necessary.

We should also try to avoid the changes to method signatures. Things like netcat_send_command should be kept with the same signature, and just update self.options['port'] (or a separate class variable self.selected_port if you think that overriding an option is a bad idea).

Regarding implementation, I am bit worried about "nc -z". This can have lots of side effect to existing services (unrelated services open). Also, sending several independent commands through cumin can be very slow. Try to send a single command checking free ports, for example, something like (as we suppose we have root rights):

netstat -altn | awk '{print $4}' | awk -F: '{print $NF}' | grep ^44[0-9][0-9]

to detect ports used in the 4400-4499 range. Feel free to improve over this one liner, I just came with this now, there could be better ways. :-)

I know, this could cause race conditions, but so do your current code, but it is a bit safer :-D. We can still do the rety (nc) once we thought the initial port was open over the list of "ports seen open", as it is unlikely ports in the 44** range will be used for other things than transfer.py. If you are still worried about race conditions, we can create a locking mechanism with a temporary dir creation, which is a transactional operation in unix (e.g. /tmp/transfer.py_port_4444).

Please also try rebasing over the new paths, that will simplify your work, I think.

Patch-set: 4
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>