Update patch set 7
Patch Set 7:

I like the port=0, it looks much nicer and retains compatibility.

Regarding tests, as this is a non-trivial change, I wonder if we should have some kind of integration tests (not unit, and not part of automatic CI) where we launch a bunch of "nc -l" and we prove it works as intended?

Patch-set: 7
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>