Update patch set 8
Patch Set 8:

(2 comments)

Please check my comments with a couple of improvements (IP hardcoding, patch summary). I will test it next.

One thing I am not happy with is mixing the commenting of unrelated portions of code AND the actual change mentioned on the patch.

IF it is not very difficult, I would split both changes on separate patches (Commenting untouched functionality of Firewall.py and MariaDB.py in one patch, and implementing the automatic detection on other).

We should keep independent changes on independent commits. I mentioned on a previous comment because a new method was being created without the comment.

It is not a big deal but will keep the patches cleaner and easier to merge. :-D

Patch-set: 8
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>