Update patch set 2
Patch Set 2:

> > The idea would be that, "if you run transfer.py --verbose" you
 > get the same output as before, but by default you get the more
 > resonable short one.
 > 
 > So should we go for two run methods? run and run_verbose?

Because all classes inherit from RemoteExecution, I think it would be cleaner if we implemented a "def __init__(self, options):" signature there (in addition to the one with no options), and we allow arbitrary options in a dictionary.

Later, on the concrete class CuminExecution, "if self.options.get('verbose', false):" we disable the verbose output.

I think that should be cleaner and will maintain compatibility with the other inherited classes.

transfer.py and Transferer.py will also need the extra verbose options.

Patch-set: 2