Update patch set 1
Patch Set 1:

(2 comments)

> I used the information from https://wikitech.wikimedia.org/wiki/Transfer.py
 > :)

That's ok. Let's make sure the defaults (when an option is omitted) are clear.

I am not sure what is the best way to maintain the wiki and this in sync. Should we output the result of --help into a text file called "README", version this and then link to this from the wiki, so we don't have to update twice?

Feel free to propose any way so you don't have to work twice :-D.

Patch-set: 1
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>