Update patch set 1
Patch Set 1:

(1 comment)

See comment below- I think it will be able to answer your question.

Please note that we want to package transfer.py (and its dependencies only), not the whole repo, which is not in a publishable state. Things like backups and other scripts will have to be installable independenctly of transfer.py, which will only be installed on our cumin hosts.

Patch-set: 1