Update patch set 1
Patch Set 1:

The test is ok, but the blind killing if the transfer fails I think is a bad idea. One of the reasons why transfer can fail is because there is other process listening on that port (transfer or otherwise).

If you intention was to kill only if the new process was confirmed as being listening, based on your comment, I am ok with it. Otherwise, I would prefer to just log an error and finish without altering the state of the target host (not try to kill what could be a different process on failure).

We can merge with just the error or do the "find process and kill it if it matches what we started" on a separate task.

Patch-set: 1
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>