Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> The test is ok, but the blind killing if the transfer fails I think is a bad idea. One of the reasons why transfer can fail is because there is other process listening on that port (transfer or otherwise).
> 
> If you intention was to kill only if the new process was confirmed as being listening, based on your comment, I am ok with it. Otherwise, I would prefer to just log an error and finish without altering the state of the target host (not try to kill what could be a different process on failure).
> 
> We can merge with just the error or do the "find process and kill it if it matches what we started" on a separate task.


I agree on with your comment, We should kill the process only if we are sure we started it. The listen job doesn't return as long as it is listening. I think this is the same issue that makes us use `time.sleep(3)` here. 

But is throwing the error necessary? Because one error like this in one machine could stop the copying of a set of other good machines too (in scenario of one source - multiple destinations). So shall we keep it as it was (we keep a record of success/failure inside the transfer_sucessful list)?

And also we can create separate task and discuss on how to go about it, what do you think?

Patch-set: 1
Reviewer: Privacybatm <8099@e9e9afe9-4712-486d-8885-f54b72dd1951>