Update patch set 8
Patch Set 8: Code-Review-1

The files under transferpy/doc/transferpy do not need to be commited in. They can be generated on the fly using sphinx-apidoc which would guarantee they are always up to date. Thus the commands would be:

 commands = sphinx-apidoc -o transferpy/doc/transferpy transferpy
            sphinx-build "transferpy/doc" "transferpy/doc/.build" --color{posargs}

The module doc can then be found at transferpy/doc/.build/transferpy/transferpy.html and is listed on the index.html page :]



Which also emits a warning about 'paramiko' missing but that is because it is not defined in the requirements.txt ;)

 WARNING: autodoc: failed to import module 'ParamikoExecution' from module 'transferpy.RemoteExecution'; the following exception was raised:
 No module named 'paramiko'

But that is unrelated to this change.

Patch-set: 8
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1