Update patch set 9
Patch Set 9:

I didn't focus too much on tox here because we are not going to setup this on CI, but the new transferpy repo once moved, and it will be fixed there once we get rid of the rest of the modules.

I just wanted to merge all open commits before migration.

Patch-set: 9