Update patch set 3
Patch Set 3:

> Could you suggest to me how? Shall we have an if-else like this:
> 
> if target.count(':') == 1:
>   path = 'point it to a default path like home? How?'

I suggest this logic:

1) Let's check that it has 1, and only 1 colon
2) Let's make sure that the substring before the colon is non-empty in both source and all targets
3) Let's make sure that the origin string after the colon is non empty

We can just error out if any of these fail, in this case hard, and tell the user to fix the parameters. Let's not try to be smart and assuming one wants a home dir because it may have bad consequences for the transfer (disk full).

Unlike the previous case, where a failure happens during transfer, I think it is ok to fail on "parameter parsing" right away.

What you did was good already, just we can do the extra 2 checks on a method with just a couple of extra lines.

Is the idea clear?

Patch-set: 3
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>