Update patch set 1
Patch Set 1:

We have not yet moved the switchover.py to transferpy. But we can move that if it is related to transferpy. In that case, we will have to move the following also:
WMFReplication
WMFMariaDB
Because those are getting imported to switchover.py

Patch-set: 1
Reviewer: Privacybatm <8099@e9e9afe9-4712-486d-8885-f54b72dd1951>