Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> (2 comments)
> 
> > Patch Set 10:
> > It is not immediately clear to me how this solved the RemoteExecution dependency for switchover.py- I don't think this ended up depending on transferpy package?
> 
> That will require code changes, which is not something i want to do in this CR. I'll handle it in a follow-up instead.

Oh, sorry, I don't mean to suggest those code changes (or at least not on this patch). So from your answer I understand that both RemoteExecution and CuminExecution will be added to the library package, correct? (it was just a question). If yes, that is the extent of my suggestions/review and this will have my +1, with some testing to do afterwards to check nothing broke after deploy-having a package is better than nothing. My general worry is that the packaging could alter some paths/dependencies and we will have to check for those after merge.

Patch-set: 11
Reviewer: Gerrit User 2556 <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>