Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> +1 if you tested this works. Only 2 questions:

I can confirm this will import the library successfully, further testing requires deploying to production.

> * Why did it use to work?

When a python script is run, python will add the current directory to the python path. When the script was in the same dir as the library, this worked fine.

> * Did you check if there is other dependencies like this?

I've had a look through all of the python files in the repo for this, i haven't found any other issues.

> I am guessing we will try to move this to transfer py dependency?

Yes.

Patch-set: 1