Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 4:
> > 
> > Should RemoteExecution be its own "module/subdir" inside this repo, given that it is use/will be used by WMFMariaDB, WMFBackup AND transfer.py, rather than being inside wmfmariadbpy?
> 
> That's something i can look into - i need to figure out how multiple top-level packages work from a setup.py perspective and from a debian packaging perspective.

Maybe let's merge as is and research on a separate patch? I remember we had issues with CI being not very friendy with with multiple submodules under certain conditions- we can have a look at it later.

The reason not to package the others is because they are interesting to keep on sources, but they have not been tested on production as much.

Patch-set: 5
Reviewer: Gerrit User 2556 <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>