Update patch set 5
Patch Set 5:

This is very much work in progress, but I wonder the performance impact of reading a file on each request. Shouldn't be as impacting given local config is also loaded + fs cache, but I don't really see a way of keeping it on memory without this being a daemon (and it is only checked once per connection).

Patch-set: 5