Update patch set 8
Patch Set 8:

> How much work is it to make the tests work under py3.5?

Could you help me find a way to mock it that is not a hack? Technically this part (WMFMariaDB and python version decision) is on your side, I am just creating the unit test for WMFMariaDB libary to prove the refactoring I need for backups works as intended.

Patch-set: 8