Update patch set 2
Patch Set 2:

How do you foresee the future development of check_private_data? So we just make changes to it on the puppet repo and ship it as it is now? 
Or should we eventually deploy it via packages?

This is not urgent and shouldn't block this patch. Do the current puppet repo version and this wmfmariadbpy version differ? (Just curious)

There might be changes needed to check_private_data.py in the future, with the new labsdb infra although I recall it is already ready to use -S $socket_location, which is all that it might need.

Patch-set: 2
Reviewer: Gerrit User 3731 <3731@e9e9afe9-4712-486d-8885-f54b72dd1951>