Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> How do you foresee the future development of check_private_data? So we just make changes to it on the puppet repo and ship it as it is now? 
> Or should we eventually deploy it via packages?

Eventually i would move it into this repo, and deploy it as a package.

> This is not urgent and shouldn't block this patch. Do the current puppet repo version and this wmfmariadbpy version differ? (Just curious)

Yes, the version in puppet has some improvements made to it (e.g. skipping comments in input files).

Currently it seems to have ~no code in common with anything else in wmfmariadbpy. I could do the migration and packaging now, but it'll eat up a bunch of time, and i'd prefer to make progress on higher priority things at the moment.

> There might be changes needed to check_private_data.py in the future, with the new labsdb infra although I recall it is already ready to use -S $socket_location, which is all that it might need.

Correct, it supports -S.

Patch-set: 2