Update patch set 3
Patch Set 3:

That is heavily tied to how the CI image spawns mysqld. The definition is at https://gerrit.wikimedia.org/r/plugins/gitiles/integration/config/+/master/dockerfiles/tox-mysqld/run-with-mysqld.sh


An alternative would be to have the test suite to spawn 1..N mysqld with controlled socket or tcp port and then refer to them in the test suite.  The advantage is one no more need a running mysqld in order to run the tests.

Patch-set: 3