resolve: Allow connections with <hostname>:<section>, in addition to port
This will allow executing things such as:

mysql.py -h db2141:s1

While:

mysql.py -h db2141:3311 still works the same.

which may be easier in some cases. This will not work if the section
name is just a number (not the case now). If the section doesn't
exist, it fails back to the standard, 3306 port.

Change-Id: I69ac313aa82a6fa9a313ee5b6d7c8040f822c7d5