Update patch set 1
Patch Set 1:

Nice! The only comment I would add is to check the md5 of the package we download from mariadb, but as this is not going to leave any test environment, I guess it is not that important?

Patch-set: 1
Reviewer: Gerrit User 3731 <3731@e9e9afe9-4712-486d-8885-f54b72dd1951>