setup.cfg: Don't specify a python_version for mypy
Recent versions of pytest and importlib-metadata for python 3.6 use
variable annotations, which isn't supported in previous python versions.

Ideally mypy would enforce python_version=3.5 for our code, and not for
the third-party dependencies. Unfortunately it does not support this.
So unless we want to start pinning all our dependencies to old versions,
the only thing we can do is drop this, and let people manually running
'tox -e py35-mypy' locally and/or CI catch these issues for us.

Example failure:
```
13:45:58 py36-mypy run-test: commands[0] | mypy .
13:45:58 [623] /src$ /src/.tox/py36/bin/mypy .
13:45:58 .tox/py36/lib/python3.6/site-packages/_pytest/assertion/__init__.py:86: error: Variable annotation syntax is only supported in Python 3.6 and greater
13:45:58 Found 1 error in 1 file (errors prevented further checking)
13:45:58 ERROR: InvocationError for command /src/.tox/py36/bin/mypy . (exited with code 2)
```

Change-Id: Ie50e8e1d36f7ae6db06b7a3fdb13aae86b06a0f8