Ensure we don't use a broken version of pip.
Pip versions 20.0.2 and earlier have a bug where the existence of a
<package>.egg-info dir in the top-level directory would cause it to
disregard develop mode. This meant that if you created a tox env without
first deleting *.egg-info, it would not do a 'develop' install, and
everything would be bad. 'wmfmariadbpy.test' would not be available
inside the venv, changes made to the working copy would not be reflected
when running tests/scripts, and life would be Sad.

https://github.com/pypa/pip/issues/7731
https://github.com/pypa/pip/issues/2926

Pip version 21.0 drops support for python 3.5, so we also add an
upper-limit constraint.

Change-Id: I0cbc00f5a4ae00ce3c91c98c894048fd43fd1164