tox/unit: Allow unit tests to be indepdenent of env vars
The current tox config will set $DBUTIL_SECTION_PORTS unless it's
running the 'unit' target. This breaks when you try to use 'cover'
target, and there doesn't seem to be a way to exclude both.

So, instead this change makes the unit tests reference the same
section_ports.csv file, and be independent of whether
$DBUTIL_SECTION_PORTS is set in the current environment or not.

Change-Id: I7799a8e22b191285a9bf4d92a81ef4fb9f6ffc73