integration: Fix master_heartbeat_period issues.
When moving a replica around, the old upstream node will have a stale
entry in its 'show slave hosts'. The duration this can persist for is
either determined by `slave_net_timeout`, or `master_heartbeat_period`,
depending on whether GTID is in use or not (apparently).

With this change we no longer need to flush any logs.

Change-Id: Ifa473249e8cab6b609be30f6c4437fcd5ec7458d