Make version numbering more Debian-standard
The previous scheme had two infelicities:

1) the main version number was like that for a pre-release:

```
matthew@tsk:~$ dpkg --compare-versions 0.12.2~wmf1 lt 0.12.2 ; echo $?
0
```
i.e. whilst we meant to have "packaging changes on top of 0.12.2", we
actually had "a version before 0.12.2". Fix that by moving to 0.12.2-1
```
matthew@tsk:~$ dpkg --compare-versions 0.12.2-1 gt 0.12.2 ; echo $?
0
```

2) the backport version number suffix +deb${DEB_VERSION_NUMBER}u1 was
greater than the main version number it was a backport from:

```
matthew@tsk:~$ dpkg --compare-versions 0.12.2~wmf1+deb12u1 gt 0.12.2~wmf1 ; echo $?
0
```

This is contrary to best practice - if you are backporting a package
from one suite to a previous, then the version of the backport must be
less than the main version, so that if the system is subsequently
upgraded to the newer suite, it will automatically pull in the
original package (rather than sticking with the backport). So use the
standard ~debXu1 convention.

```
matthew@tsk:~$ dpkg --compare-versions 0.12.2-1~deb12u1 lt 0.12.2-1 ; echo $?
0
```

While I'm here, also adjust the .backport job a little so the tasks
that use it can be a little more DRY.